+++ data-updates.lua
@@ -1051,4 +1051,5 @@ end
 
 fw.start_converting()
 
--- error("STOP")
+log("!!! - " .. serpent.block(data.raw["recipe"]["sulfuric-acid"]))
+--error("STOP")